requirements.txt
absl-py==1.1.0
asgiref==3.5.2
astunparse==1.6.3
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
colour==0.1.5
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
docopt==0.6.2
filelock==3.7.1
flatbuffers==1.12
gast==0.4.0
google-api-core==2.8.2
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-cloud-vision==2.7.3
google-pasta==0.2.0
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
h5py==3.7.0
huggingface-hub==0.7.0
idna==3.3
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
nltk==3.7
numpy==1.22.4
oauthlib==3.2.0
odfpy==1.4.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pdf2image==1.16.0
pdfminer==20191125
pdfminer.six==20220524
pdfminer3k==1.3.4
pdfplumber==0.7.1
Pillow==9.1.1
pipreq==0.4
pipreqs==0.4.11
ply==3.11
proto-plus==1.20.6
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pyparsing==3.0.9
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
sentence-transformers==2.2.2
sentencepiece==0.1.96
six==1.16.0
slate3k==0.5.3
soupsieve==2.3.2.post1
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.20.0
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
Wand==0.6.7
Werkzeug==2.1.2
wrapt==1.14.1
yarg==0.1.9
